Moved validation to Plugin manager and added controll if there is any dependecy
